+++ tests/fal-webhook-architecture.test.ts
@@ -1,0 +1,74 @@
+import assert from 'node:assert/strict';
+import { existsSync, readFileSync } from 'node:fs';
+import { join } from 'node:path';
+import test from 'node:test';
+
+const root = process.cwd();
+const handlerPath = join(root, 'frontend/server/fal-webhook-handler.ts');
+const mappingPath = join(root, 'frontend/server/fal-webhook-mapping.ts');
+
+const handlerSource = readFileSync(handlerPath, 'utf8');
+const mappingSource = readFileSync(mappingPath, 'utf8');
+
+test('Fal webhook handler delegates mapping and payload extraction helpers', () => {
+  assert.ok(existsSync(mappingPath), 'Fal webhook mapping helpers should live in a sibling server module');
+  assert.match(handlerSource, /from '\.\/fal-webhook-mapping'/, 'webhook handler should import mapping helpers');
+
+  for (const implementationName of [
+    'fallbackThumbnail',
+    'formatAspectRatioLabel',
+    'extractStringField',
+    'extractIdentifiersFromPayload',
+    'inferEngineFromPayload',
+    'findFirstString',
+    'extractMediaUrls',
+    'normalizeRenderIdList',
+    'extractImageUrlsFromPayload',
+    'normalizeErrorText',
+    'findFirstErrorMessage',
+    'extractFalErrorMessage',
+    'normalizeStatus',
+  ]) {
+    assert.doesNotMatch(
+      handlerSource,
+      new RegExp(`function ${implementationName}\\(`),
+      `${implementationName} belongs in fal-webhook-mapping.ts`
+    );
+  }
+
+  assert.doesNotMatch(handlerSource, /PROVIDER_ENGINE_MAP/, 'provider engine mapping belongs in fal-webhook-mapping.ts');
+  assert.doesNotMatch(handlerSource, /ERROR_MESSAGE_KEYS/, 'error message key mapping belongs in fal-webhook-mapping.ts');
+  assert.doesNotMatch(handlerSource, /listUpscaleToolEngines/, 'tool engine media lookup belongs in fal-webhook-mapping.ts');
+
+  const lineCount = handlerSource.split('\n').length;
+  assert.ok(lineCount <= 950, `Fal webhook handler should stay below 950 lines after mapping extraction, got ${lineCount}`);
+});
+
+test('Fal webhook mapping module exposes the expected helper contract', () => {
+  for (const exportName of [
+    'fallbackThumbnail',
+    'formatAspectRatioLabel',
+    'extractIdentifiersFromPayload',
+    'inferEngineFromPayload',
+    'findFirstString',
+    'extractMediaUrls',
+    'normalizeRenderIdList',
+    'extractImageUrlsFromPayload',
+    'extractFalErrorMessage',
+    'normalizeStatus',
+    'getUpscaleToolMediaType',
+    'isCompletedFalStatus',
+    'isFailedFalStatus',
+  ]) {
+    assert.match(
+      mappingSource,
+      new RegExp(`export (async function|function) ${exportName}`),
+      `${exportName} should be exported by fal-webhook-mapping.ts`
+    );
+  }
+
+  assert.match(mappingSource, /export type FalWebhookPayload =/, 'FalWebhookPayload should move with mapping helpers');
+  assert.match(mappingSource, /export type WebhookIdentifiers =/, 'WebhookIdentifiers should move with mapping helpers');
+  assert.match(mappingSource, /const PROVIDER_ENGINE_MAP/, 'provider alias map should live with engine inference');
+  assert.match(mappingSource, /const ERROR_MESSAGE_KEYS/, 'error message key map should live with error extraction');
+});
